requirements.txt
python-dotenv
sendgrid
pika
pandas
xlsxwriter
aio-pika
certifi